docs: fix codeblock in FAQ section of readme (#36) (#50)
On w.org code-blocks in FAQ are not rendered properly. Using single
backticks does not give the best experience in rendered MarkDown, e.g.
on GitHub, but it's still readable and it's correctly rendered on the
plugin page.

We might want to migrate to a dedicated readme.txt for deployment in the
future, but for now let's use the cheap solution so we are not blocked.